[Chore] #50 - ShareCard의 Header를 별도의 파일로 분리
